Fix synchronization problem
The scheduler pulls data off the queue
on a different thread than a call to the client#join
method. While the connector is busy writing data,
the queue is sitting empty. Join sees no data left
on the queue and allows the program to terminate,
even though the connector has not finished.